fatigue: goodman_haigh() moved from module 'sn' to 'corrections'
